third_party/WebKit/LayoutTests/editing/pasteboard/copy-paste-float-expected.txt

CONSOLE MESSAGE: line 18: SUCCESS: paste DID NOT keep float:left style.
This test is to see if text within a floating block is put onto the paste board wihtout the float style.

First LI
Second LI

---

First LI
Second LI
---
